добавил подходы в сессию + защитил все юзкейсы проверками экзист на юзера/др. сущности , спасая от эрор500

+++ src/main/java/com/example/fitness_assistant/application/service/workout/DeleteWorkoutUseCase.java
@@ -1,6 +1,8 @@
 package com.example.fitness_assistant.application.service.workout;
 
-import com.example.fitness_assistant.core.exception.HydrationNotFoundException;
+import com.example.fitness_assistant.core.exception.UserNotFoundException;
+import com.example.fitness_assistant.core.exception.WorkoutNotFoundException;
+import com.example.fitness_assistant.core.repository.UserRepository;
 import com.example.fitness_assistant.core.repository.WorkoutRepository;
 import com.example.fitness_assistant.infrastructure.security.UserDetailsAdapter;
 import lombok.RequiredArgsConstructor;
@@ -12,13 +14,17 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class DeleteWorkoutUseCase {
 
+    private final UserRepository userRepository;
     private final WorkoutRepository workoutRepository;
 
     @Transactional
     public void deleteWorkout(Long id, UserDetails userDetails) {
         Long userId = ((UserDetailsAdapter) userDetails).getUserId();
+        if (!userRepository.existsById(userId)) {
+            throw new UserNotFoundException(userId);
+        }
         if (!workoutRepository.existsById(id, userId)) {
-            throw new HydrationNotFoundException(id);
+            throw new WorkoutNotFoundException(id);
         }
         workoutRepository.deleteById(id, userId);
     }
